Refactored routes and requests, now don't need to pass userId anymore

+++ src/main/kotlin/routes/PostRoutes.kt
@@ -5,20 +5,17 @@ import com.example.data.requests.DeletePostRequest
 import com.example.data.response.BasicApiResponse
 import com.example.service.LikeService
 import com.example.service.PostService
-import com.example.service.UserService
 import com.example.util.ApiResponseMessage
 import com.example.util.Constants
 import com.example.util.QueryParams
 import io.ktor.http.*
 import io.ktor.server.auth.*
-import io.ktor.server.auth.jwt.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 fun Route.createPost(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         post("/api/post/create") {
@@ -30,20 +27,7 @@ fun Route.createPost(
                 return@post
             }
 
-            val email = call.principal<JWTPrincipal>()?.getClaim("email", String:: class)
-            val isEmailByUser = userService.doseEmailBelongToUserId(
-                email = email ?: "",
-                userId = request.userId
-            )
-            if (!isEmailByUser) {
-                call.respond(
-                    status = HttpStatusCode.Unauthorized,
-                    message = "You are not who you say you are."
-                )
-                return@post
-            }
-
-            val didUserExist = postService.createPostIfUserExist(request)
+            val didUserExist = postService.createPostIfUserExist(request, call.userId)
             if (!didUserExist) {
                 call.respond(
                     status = HttpStatusCode.OK,
@@ -66,30 +50,13 @@ fun Route.createPost(
 
 fun Route.getPostForFollows(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         get("api/post/get") {
-            val userId = call.parameters[QueryParams.PARAM_USER_ID] ?: kotlin.run {
-                call.respond(HttpStatusCode.BadRequest)
-                return@get
-            }
             val page = call.parameters[QueryParams.PARAM_PAGE]?.toIntOrNull() ?: 0
             val pageSize = call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toIntOrNull() ?: Constants.DEFAULT_POST_PAGE_SIZE
 
-            val email = call.principal<JWTPrincipal>()?.getClaim("email", String:: class)
-            val isEmailByUser = userService.doseEmailBelongToUserId(
-                email = email ?: "",
-                userId = userId
-            )
-            if (!isEmailByUser) {
-                call.respond(
-                    status = HttpStatusCode.Unauthorized,
-                    message = "You are not who you say you are."
-                )
-                return@get
-            }
-            val posts = postService.getPostsForFollows(userId, page, pageSize)
+            val posts = postService.getPostsForFollows(call.userId, page, pageSize)
             call.respond(
                 HttpStatusCode.OK,
                 posts
@@ -116,16 +83,20 @@ fun Route.deletePost(
                 call.respond(HttpStatusCode.NotFound)
                 return@delete
             }
-            postService.deletePost(request.postId)
-            likeService.deleteLikesForParent(request.postId)
-            call.respond(
-                status = HttpStatusCode.OK,
-                message = BasicApiResponse(
-                    successful = true,
-                    message = "Post deleted Successfully."
-                )
-            )
 
+            if (post.userId == call.userId) {
+                postService.deletePost(request.postId)
+                likeService.deleteLikesForParent(request.postId)
+                call.respond(
+                    status = HttpStatusCode.OK,
+                    message = BasicApiResponse(
+                        successful = true,
+                        message = "Post deleted Successfully."
+                    )
+                )
+            } else {
+                call.respond(HttpStatusCode.Unauthorized)
+            }
         }
     }
 }
